[Decomp][Spell] Cohesion split: extract SpellEffectDamageTeleport handlers to SpellEffectDamageTeleport.cpp
